+++ src/extension.ts
@@ -8,9 +8,7 @@
 // ------------------------------------------------------------------------
 
 import * as vscode from 'vscode';
-import * as transformutilities from './common/transformutilities';
-import * as lineutilities from './common/lineutilities';
-import * as insertutilities from './common/insertutilities';
+import * as utilityList from './common/utility-list';
 
 // ------------------------------------------------------------------------
 //                               Startup Code
@@ -20,40 +18,7 @@ import * as insertutilities from './common/insertutilities';
 // your extension is activated the very first time the command is executed
 export function activate(context: vscode.ExtensionContext) {
 
-    interface TUtilityDef {
-        f: () => void;
-        id: string;
-    }
-
-    const utilities: TUtilityDef[] = [
-        { f: transformutilities.capitalize, id: 'editor.capitalize' },
-        { f: transformutilities.camelCase, id: 'editor.camelCase' },
-        { f: transformutilities.spaceByUpper, id: 'editor.spaceByUpper' },
-        { f: transformutilities.reverseAssignment, id: 'editor.reverseAssignment' },
-        { f: transformutilities.unixToWinSlash, id: 'editor.unixToWinSlash' },
-        { f: transformutilities.winToUnixSlash, id: 'editor.winToUnixSlash' },
-        { f: transformutilities.singleToDoubleSlash, id: 'editor.singleToDoubleSlash' },
-        { f: transformutilities.doubleToSingleSlash, id: 'editor.doubleToSingleSlash' },
-        { f: transformutilities.urlEncode, id: 'editor.urlEncode' },
-        { f: transformutilities.urlDecode, id: 'editor.urlDecode' },
-        { f: transformutilities.regnize, id: 'editor.regnize' },
-        { f: transformutilities.headerToBookmark, id: 'editor.headerToBookmark' },
-        { f: transformutilities.mixer, id: 'editor.mixer' },
-        { f: lineutilities.removeDuplicatedLines, id: 'editor.removeDuplicatedLines' },
-        { f: lineutilities.removeEmptyLines, id: 'editor.removeEmptyLines' },
-        { f: lineutilities.joinLines, id: 'editor.joinLines' },
-        { f: lineutilities.splitLines, id: 'editor.splitLines' },
-        { f: lineutilities.sortNumericallyAscending, id: 'editor.sortNumericallyAscending' },
-        { f: lineutilities.indentOneSpace, id: 'editor.indentOneSpace' },
-        { f: lineutilities.outdentOneSpace, id: 'editor.outdentOneSpace' },
-        { f: insertutilities.insertISODate, id: 'editor.insertISODate' },
-        { f: insertutilities.insertISOTimeDate, id: 'editor.insertISOTimeDate' },
-        { f: insertutilities.insertUUID, id: 'editor.insertUUID' },
-        { f: insertutilities.insertTextAtEnd, id: 'editor.insertTextAtEnd' },
-        { f: insertutilities.insertTextAtStart, id: 'editor.insertTextAtStart' },
-    ];
-
-    for (const action of utilities) {
+    for (const action of utilityList.utilityList) {
         const disposable = vscode.commands.registerCommand(action.id, action.f);
         context.subscriptions.push(disposable);
     }